refactor: simplify discriminators directory structure
