+++ src/discriminators/__mocks__/droidSchema.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { Schema } from './mongooseCommon';
+import { Schema } from '../../__mocks__/mongooseCommon';
 
 export const DroidSchema = new Schema({
   makeDate: Date,
