+++ src/discriminators/__mocks__/personSchema.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import { Schema } from './mongooseCommon';
+import { Schema } from '../../__mocks__/mongooseCommon';
 
 export const PersonSchema = new Schema({
   dob: Number,
