+++ src/discriminators/__tests__/prepareBaseResolvers-test.js
@@ -1,14 +1,14 @@
 /* @flow */
 
 import { graphql } from 'graphql-compose';
-import { getCharacterModels } from '../../../__mocks__/characterModels';
-import { composeWithMongooseDiscriminators } from '../../../composeWithMongooseDiscriminators';
+import { getCharacterModels } from '../__mocks__/characterModels';
+import { composeWithMongooseDiscriminators } from '../../composeWithMongooseDiscriminators';
 
 const { CharacterModel } = getCharacterModels('type');
 
 const CharacterDTC = composeWithMongooseDiscriminators(CharacterModel);
 
-describe('recomposeBaseResolvers()', () => {
+describe('prepareBaseResolvers()', () => {
   describe('setDKeyEnumOnITCArgs()', () => {
     const resolversWithFilterAndRecordArgs = [];
     const resolversWithFilterArgsOnly = [];
