+++ src/discriminators/prepareBaseResolvers.js
@@ -1,8 +1,8 @@
 /* @flow */
 
 import { graphql } from 'graphql-compose';
-import { DiscriminatorTypeComposer } from '../DiscriminatorTypeComposer';
-import { EMCResolvers } from '../../resolvers';
+import { DiscriminatorTypeComposer } from './DiscriminatorTypeComposer';
+import { EMCResolvers } from '../resolvers';
 
 const { GraphQLList, GraphQLNonNull } = graphql;
