+++ src/discriminators/utils/__test__/mergeTypeConverterResolverOpts-test.js
@@ -1,6 +1,6 @@
 /* @flow */
 
-import type { TypeConverterResolversOpts } from '../../../../composeWithMongoose';
+import type { TypeConverterResolversOpts } from '../../../composeWithMongoose';
 import {
   mergeTypeConverterResolverOpts,
   mergePrimitiveTypeFields,
