--- src/discriminators/utils/index.js
@@ -1,3 +1,0 @@
-/* @flow */
-
-export { reorderFields } from './reorderFields';
